requirements.txt
python-jose[cryptography]
passlib[bcrypt]
fastapi
uvicorn
python-multipart
psycopg2-binary